Removes all of the airbnb lint warnings/errors
