+++ star-button.js
@@ -1,0 +1,74 @@
+// React and react native imports
+import React, {
+  Component,
+  PropTypes,
+} from 'react';
+
+// Third party imports
+import Button from 'react-native-button';
+import EntypoIcons from 'react-native-vector-icons/Entypo';
+import EvilIconsIcons from 'react-native-vector-icons/EvilIcons';
+import FontAwesomeIcons from 'react-native-vector-icons/FontAwesome';
+import FoundationIcons from 'react-native-vector-icons/Foundation';
+import IoniconsIcons from 'react-native-vector-icons/Ionicons';
+import MaterialIconsIcons from 'react-native-vector-icons/MaterialIcons';
+import OcticonsIcons from 'react-native-vector-icons/Octicons';
+import ZocialIcons from 'react-native-vector-icons/Zocial';
+
+const iconSets = {
+  Entypo: EntypoIcons,
+  EvilIcons: EvilIconsIcons,
+  FontAwesome: FontAwesomeIcons,
+  Foundation: FoundationIcons,
+  Ionicons: IoniconsIcons,
+  MaterialIcons: MaterialIconsIcons,
+  Octicons: OcticonsIcons,
+  Zocial: ZocialIcons,
+};
+
+class StarButton extends Component {
+
+  constructor(props) {
+    super(props);
+
+    this.onButtonPress = this.onButtonPress.bind(this);
+  }
+
+  onButtonPress() {
+    this.props.onStarButtonPress(this.props.rating);
+  }
+
+  render() {
+    const Icon = iconSets[this.props.iconSet];
+
+    return (
+      <Button
+        activeOpacity={0.20}
+        disabled={this.props.disabled}
+        onPress={this.onButtonPress}
+        style={{
+          height: this.props.starSize,
+          width: this.props.starSize,
+        }}
+      >
+        <Icon
+          name={this.props.starIconName}
+          size={this.props.starSize}
+          color={this.props.starColor}
+        />
+      </Button>
+    );
+  }
+}
+
+StarButton.propTypes = {
+  disabled: PropTypes.bool,
+  rating: PropTypes.number,
+  onStarButtonPress: PropTypes.func,
+  iconSet: PropTypes.string,
+  starSize: PropTypes.number,
+  starIconName: PropTypes.string,
+  starColor: PropTypes.string,
+};
+
+export default StarButton;
